% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adstock.R
\name{adstock}
\alias{adstock}
\title{Create adstock function to use for media retention}
\usage{
adstock(retention)
}
\arguments{
\item{retention}{is a required numeric value to return the appropriate
adstock function which can range from 0 to 1.}
}
\value{
A function is returned from a function factory after a value is
specified in \code{retention} which can only range from 0 to 1.
}
\description{
Create adstock function to use for media retention
}
\details{
Advertising adstock or advertising carry-over is the prolonged or
lagged effect of advertising. adstock returns a function with the required
retention factor built into it depending on specified value by the user which
can range from 0 to 1.
input variable
}
