% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_analysis.R
\docType{package}
\name{CampaignAnalysis}
\alias{CampaignAnalysis}
\alias{CampaignAnalysis-package}
\title{Io output from ebuilder express}
\description{
This package contains all the calculations shown in IO platform.
}
