% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{mixed_linear_model}
\alias{mixed_linear_model}
\title{Mixed effects with}
\usage{
mixed_linear_model(database)
}
\arguments{
\item{database}{The base to transform}
}
\value{
A data.table object
}
\description{
Transforms factor variables into character or numeric variables.
}
